voyage/pyproject.toml
[project]
name = "voyage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "voyageai",
    "Pillow",
    "pymongo"
]
